moved Item entities out of holdings file and into their own file
