+++ spec/spectrum/presenters/holdings_presenter_spec.rb
@@ -1,4 +1,4 @@
-require_relative '../../../spec_helper'
+require_relative '../../spec_helper'
 
 describe Spectrum::Presenters::HathiTrustHoldingPresenter, "to_h" do
   before(:each) do
