+++ spec/spectrum/presenters/mirlyn_item_presenter_spec.rb
@@ -1,4 +1,4 @@
-require_relative '../../../spec_helper'
+require_relative '../../spec_helper'
 
 describe Spectrum::Presenters::MirlynItem, "to_a" do
     before(:each) do
